feat: unstack media modals
Currently, in the EditPage layout, when you try to upload a photo,
a modal pops up within a modal - this is bad UI which might confuse
the users. The proposed solution is to close one modal when another one
is opened, that is, when someone clicks on "Upload photo", the original
media modal closes and the media setting modal opens. To do this, state
needs to be maintained on the parent EditPage layout instead of on the
media modal. This commit implements that fix.

Note: this change needs to be similarly applied to the EditResourcesPage
and EditCollectionPage as well.
